created simple workflows, add .gitignore

+++ conv_helpers.c
@@ -13,7 +13,8 @@ static size_t print_conv_bin(char *outbuf, char *inbuf, size_t inbuf_size)
 	for (i = 0; i < inbuf_size; i++) {
 		byte = inbuf[i];
 		if (i % 4 == 0)
-			p += scnprintf(p, OFFSET_STR_BUF, "0x%04x ", i);
+			p += scnprintf(p, OFFSET_STR_BUF, "0x%04x ",
+					(unsigned int)i);
 		for (shift = 7; shift >= 0; shift--) {
 			*p++ = (byte & (1 << shift)) ? '1' : '0';
 			if (shift == 4) *p++ = ' ';
@@ -40,7 +41,8 @@ static size_t print_conv_octal(char *outbuf, char *inbuf, size_t inbuf_size)
 	for (i = 0; i < (ssize_t)inbuf_size; i++) {
 		byte = inbuf[i];
 		if (i % 8 == 0)
-			p += scnprintf(p, OFFSET_STR_BUF, "0x%04x ", i);
+			p += scnprintf(p, OFFSET_STR_BUF, "0x%04x ",
+					(unsigned int)i);
 		*p++ = 'o';
 		for (shift = 6; shift >= 0; shift -= 3)
 			*p++ = ((byte >> shift) & 7) + '0';
